+++ cruise-control/src/main/java/com/linkedin/kafka/cruisecontrol/detector/KafkaMetricAnomaly.java
@@ -26,7 +26,7 @@ public class KafkaMetricAnomaly implements MetricAnomaly<BrokerEntity> {
   private final KafkaCruiseControl _kafkaCruiseControl;
   private final String _description;
   private final BrokerEntity _brokerEntity;
-  private final Integer _metricId;
+  private final Short _metricId;
   private final List<Long> _windows;
   private final String _anomalyId;
   private OptimizationResult _optimizationResult;
@@ -43,7 +43,7 @@ public class KafkaMetricAnomaly implements MetricAnomaly<BrokerEntity> {
   public KafkaMetricAnomaly(KafkaCruiseControl kafkaCruiseControl,
                             String description,
                             BrokerEntity brokerEntity,
-                            Integer metricId,
+                            Short metricId,
                             List<Long> windows) {
     _kafkaCruiseControl = kafkaCruiseControl;
     _description = description;
@@ -82,7 +82,7 @@ public class KafkaMetricAnomaly implements MetricAnomaly<BrokerEntity> {
    * Get the metric Id caused the metric anomaly.
    */
   @Override
-  public Integer metricId() {
+  public Short metricId() {
     return _metricId;
   }
